+++ contracts/btc-finality/src/contract.rs
@@ -8,20 +8,17 @@ use crate::state::config::{
     Config, ADMIN, CONFIG, DEFAULT_JAIL_DURATION, DEFAULT_MAX_ACTIVE_FINALITY_PROVIDERS,
     DEFAULT_MIN_PUB_RAND, DEFAULT_MISSED_BLOCKS_WINDOW, DEFAULT_REWARD_INTERVAL,
 };
-
+use crate::state::finality::get_btc_staking_activated_height;
 use crate::{finality, queries, state};
 
 #[cfg(test)]
 use crate::state::finality::ACCUMULATED_VOTING_WEIGHTS;
 use babylon_apis::finality_api::SudoMsg;
-
-use btc_staking::msg::ActivatedHeightResponse;
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
 use cosmwasm_std::{
     attr, coin, to_json_binary, Addr, CustomQuery, Deps, DepsMut, Empty, Env, MessageInfo,
-    QuerierWrapper, QueryRequest, QueryResponse, Reply, Response, StdResult, Uint128, WasmMsg,
-    WasmQuery,
+    QueryRequest, QueryResponse, Reply, Response, StdResult, Uint128, WasmMsg, WasmQuery,
 };
 use cw2::set_contract_version;
 use cw_utils::{maybe_addr, nonpayable};
@@ -125,6 +122,14 @@ pub fn query(deps: Deps, _env: Env, msg: QueryMsg) -> Result<QueryResponse, Cont
         QueryMsg::Votes { height } => Ok(to_json_binary(&queries::votes(deps, height)?)?),
         QueryMsg::SigningInfo { btc_pk_hex } => {
             Ok(to_json_binary(&queries::signing_info(deps, btc_pk_hex)?)?)
+        }
+        QueryMsg::ActivatedHeight {} => {
+            let activated_height = get_btc_staking_activated_height(deps.storage);
+            if let Some(height) = activated_height {
+                Ok(to_json_binary(&height)?)
+            } else {
+                Err(ContractError::BTCStakingNotActivated)
+            }
         }
     }
 }
@@ -238,17 +243,25 @@ fn handle_end_block(
     // finality provider has voting power, start indexing and tallying blocks
     let cfg = CONFIG.load(deps.storage)?;
     let mut res = Response::new();
-    let activated_height = get_activated_height(&cfg.staking, &deps.querier)?;
-    if activated_height > 0 {
-        // Index the current block
-        let ev = finality::index_block(deps, env.block.height, &hex::decode(app_hash_hex)?)?;
-        res = res.add_event(ev);
-        // Tally all non-finalised blocks
-        let events = finality::tally_blocks(deps, &env, activated_height)?;
-        res = res.add_events(events);
+
+    let activated_height = get_btc_staking_activated_height(deps.storage);
+
+    // If the BTC staking protocol is not activated, do nothing
+    if activated_height.is_none() {
+        return Ok(res);
     }
 
-    // On a reward distribution boundary, calculate and send rewards for distribution to Babylon Genesis
+    let activated_height = activated_height.unwrap();
+
+    // Index the current block
+    let ev = finality::index_block(deps, env.block.height, &hex::decode(app_hash_hex)?)?;
+    res = res.add_event(ev);
+
+    // Tally all non-finalised blocks
+    let events = finality::tally_blocks(deps, &env, activated_height)?;
+    res = res.add_events(events);
+
+    // On an reward interval boundary, send rewards for distribution to Babylon Genesis
     if env.block.height > 0 && env.block.height % cfg.reward_interval == 0 {
         if let Some(rewards_msg) = handle_rewards_distribution(deps, &env)? {
             res = res.add_message(rewards_msg);
@@ -256,16 +269,6 @@ fn handle_end_block(
     }
 
     Ok(res)
-}
-
-pub fn get_activated_height(staking_addr: &Addr, querier: &QuerierWrapper) -> StdResult<u64> {
-    // TODO: Use a raw query (#41)
-    let query = encode_smart_query(
-        staking_addr,
-        &btc_staking::msg::QueryMsg::ActivatedHeight {},
-    )?;
-    let res: ActivatedHeightResponse = querier.query(&query)?;
-    Ok(res.height)
 }
 
 pub(crate) fn encode_smart_query<Q: CustomQuery>(